# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: BFF OpenAPI Support User
  version: 1.0.0
  description: API for user management.
  contact:
    name: Banco Fortaleza Financiera

servers:
  - url: /
    description: Default server.

tags:
  - name: Users
    description: Operations related to users.
  - name: UserAddresses
    description: Operations related to user addresses.
  - name: UserPhones
    description: Operations related to user phones.

paths:
  /support/v1/users:
    get:
      tags:
        - Users
      summary: List users
      description: Retrieves users using the submitted filters when provided.
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/XPageHeader'
        - $ref: '#/components/parameters/XPageSizeHeader'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/UserTypeQuery'
      responses:
        '200':
          description: Users retrieved successfully.
          headers:
            x-total-count:
              $ref: '#/components/headers/XTotalCount'
            x-page:
              $ref: '#/components/headers/XPage'
            x-page-size:
              $ref: '#/components/headers/XPageSize'
            x-total-pages:
              $ref: '#/components/headers/XTotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Users
      summary: Create user
      description: Creates a user with personal information, credentials, user type, status and optional contact data.
      operationId: createUser
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /support/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by id
      description: Retrieves the user that matches the submitted unique identifier.
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - Users
      summary: Update user status
      description: Updates the user status. Use INACTIVE to deactivate a user.
      operationId: updateUserStatus
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          description: User status updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /support/v1/users/{id}/addresses:
    get:
      tags:
        - UserAddresses
      summary: List user addresses
      description: Retrieves addresses associated with the submitted user identifier using the submitted filters when provided.
      operationId: listUserAddresses
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/XPageHeader'
        - $ref: '#/components/parameters/XPageSizeHeader'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/StatusQuery'
      responses:
        '200':
          description: User addresses retrieved successfully.
          headers:
            x-total-count:
              $ref: '#/components/headers/XTotalCount'
            x-page:
              $ref: '#/components/headers/XPage'
            x-page-size:
              $ref: '#/components/headers/XPageSize'
            x-total-pages:
              $ref: '#/components/headers/XTotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - UserAddresses
      summary: Create user address
      description: Creates an address associated with the submitted user identifier.
      operationId: createUserAddress
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressCreateRequest'
      responses:
        '201':
          description: User address created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /support/v1/users/{id}/addresses/{addressId}:
    get:
      tags:
        - UserAddresses
      summary: Get user address by id
      description: Retrieves the address that matches the submitted user and address identifiers.
      operationId: getUserAddressById
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/AddressId'
      responses:
        '200':
          description: User address retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - UserAddresses
      summary: Update user address status
      description: Updates the user address status. Use INACTIVE to deactivate an address.
      operationId: updateUserAddressStatus
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/AddressId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          description: User address status updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /support/v1/users/{id}/phones:
    get:
      tags:
        - UserPhones
      summary: List user phones
      description: Retrieves phones associated with the submitted user identifier using the submitted filters when provided.
      operationId: listUserPhones
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/XPageHeader'
        - $ref: '#/components/parameters/XPageSizeHeader'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/StatusQuery'
      responses:
        '200':
          description: User phones retrieved successfully.
          headers:
            x-total-count:
              $ref: '#/components/headers/XTotalCount'
            x-page:
              $ref: '#/components/headers/XPage'
            x-page-size:
              $ref: '#/components/headers/XPageSize'
            x-total-pages:
              $ref: '#/components/headers/XTotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhoneResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - UserPhones
      summary: Create user phone
      description: Creates a phone associated with the submitted user identifier.
      operationId: createUserPhone
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneCreateRequest'
      responses:
        '201':
          description: User phone created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /support/v1/users/{id}/phones/{phoneId}:
    get:
      tags:
        - UserPhones
      summary: Get user phone by id
      description: Retrieves the phone that matches the submitted user and phone identifiers.
      operationId: getUserPhoneById
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/PhoneId'
      responses:
        '200':
          description: User phone retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags:
        - UserPhones
      summary: Update user phone status
      description: Updates the user phone status. Use INACTIVE to deactivate a phone.
      operationId: updateUserPhoneStatus
      parameters:
        - $ref: '#/components/parameters/XDeviceIpHeader'
        - $ref: '#/components/parameters/XSessionHeader'
        - $ref: '#/components/parameters/XUserIdHeader'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/PhoneId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          description: User phone status updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    XDeviceIpHeader:
      name: x-device-ip
      in: header
      required: true
      description: Client device IP address.
      schema:
        type: string
        minLength: 7
        maxLength: 45
        pattern: '^(((25[0-5]|2[0-4][0-9]|1?[0-9]{1,2})\.){3}(25[0-5]|2[0-4][0-9]|1?[0-9]{1,2})|([0-9a-fA-F]{0,4}:){2,7}[0-9a-fA-F]{0,4})$'
        example: "192.168.1.10"
    XSessionHeader:
      name: x-session
      in: header
      required: true
      description: Client session identifier.
      schema:
        type: string
        minLength: 36
        maxLength: 36
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        example: "7f2c1a54-4cf0-4d8c-9b18-2a6a57e7f7f3"
    XUserIdHeader:
      name: x-userid
      in: header
      required: true
      description: Authenticated user identifier.
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 10
    XPageHeader:
      name: x-page
      in: header
      required: false
      description: Page number to retrieve.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
        example: 1
    XPageSizeHeader:
      name: x-page-size
      in: header
      required: false
      description: Maximum number of resources to retrieve per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    SearchQuery:
      name: search
      in: query
      required: false
      description: Search value used to find partial matches in the backend-defined searchable fields for this resource.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        example: "Gerson"
    UserId:
      name: id
      in: path
      required: true
      description: Unique user identifier.
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 1
    AddressId:
      name: addressId
      in: path
      required: true
      description: Unique address identifier.
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 1
    PhoneId:
      name: phoneId
      in: path
      required: true
      description: Unique phone identifier.
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 1
    StatusQuery:
      name: status
      in: query
      required: false
      description: Filters resources by status.
      schema:
        $ref: '#/components/schemas/Status'
    UserTypeQuery:
      name: userType
      in: query
      required: false
      description: Filters users by user type.
      schema:
        $ref: '#/components/schemas/UserType'

  headers:
    XTotalCount:
      description: Total number of resources matching the submitted filters.
      schema:
        type: integer
        format: int32
        minimum: 0
        example: 125
    XPage:
      description: Current page number returned in the response.
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 1
    XPageSize:
      description: Maximum number of resources returned per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        example: 20
    XTotalPages:
      description: Total number of pages available for the submitted filters.
      schema:
        type: integer
        format: int32
        minimum: 0
        example: 7

  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MethodNotAllowed:
      description: Method not allowed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Status:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
      minLength: 6
      maxLength: 8
      example: ACTIVE

    UserType:
      type: string
      enum:
        - ADMIN
        - NORMAL
      minLength: 5
      maxLength: 6
      example: NORMAL

    UserCreateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
          example: "Gerson"
        lastName:
          type: string
          minLength: 1
          maxLength: 64
          example: "Ramos"
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          writeOnly: true
          example: "Str0ngP@ssword"
        status:
          $ref: '#/components/schemas/Status'
        userType:
          $ref: '#/components/schemas/UserType'
        addresses:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/AddressCreateRequest'
        phones:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/PhoneCreateRequest'
      required:
        - name
        - lastName
        - password

    StatusUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        status:
          $ref: '#/components/schemas/Status'
      required:
        - status

    UserResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          readOnly: true
          example: 1
        name:
          type: string
          minLength: 1
          maxLength: 64
          example: "Gerson"
        lastName:
          type: string
          minLength: 1
          maxLength: 64
          example: "Ramos"
        status:
          $ref: '#/components/schemas/Status'
        userType:
          $ref: '#/components/schemas/UserType'
        addresses:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/AddressResponse'
        phones:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/PhoneResponse'
        created_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          example: "2026-05-01T16:35:00Z"
        updated_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          example: "2026-05-01T16:40:00Z"
      required:
        - id
        - name
        - lastName
        - status
        - userType

    AddressCreateRequest:
      type: object
      additionalProperties: false
      properties:
        address:
          type: string
          minLength: 1
          maxLength: 128
          example: "10 Calle 1-23 Zona 10"
        typeAddress:
          type: string
          minLength: 1
          maxLength: 32
          example: "HOME"
        status:
          $ref: '#/components/schemas/Status'
      required:
        - address
        - typeAddress

    AddressResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          readOnly: true
          example: 1
        idUser:
          type: integer
          format: int32
          minimum: 1
          readOnly: true
          example: 1
        address:
          type: string
          minLength: 1
          maxLength: 128
          example: "10 Calle 1-23 Zona 10"
        typeAddress:
          type: string
          minLength: 1
          maxLength: 32
          example: "HOME"
        status:
          $ref: '#/components/schemas/Status'
        created_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          example: "2026-05-01T16:35:00Z"
        updated_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          example: "2026-05-01T16:40:00Z"
      required:
        - id
        - idUser
        - address
        - typeAddress
        - status

    PhoneCreateRequest:
      type: object
      additionalProperties: false
      properties:
        phone:
          type: string
          minLength: 1
          maxLength: 16
          pattern: '^[0-9+ -]{1,16}$'
          example: "+502 5555 1234"
        typePhone:
          type: string
          minLength: 1
          maxLength: 32
          example: "MOBILE"
        status:
          $ref: '#/components/schemas/Status'
      required:
        - phone
        - typePhone

    PhoneResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
          readOnly: true
          example: 1
        idUser:
          type: integer
          format: int32
          minimum: 1
          nullable: true
          readOnly: true
          example: 1
        phone:
          type: string
          minLength: 1
          maxLength: 16
          pattern: '^[0-9+ -]{1,16}$'
          example: "+502 5555 1234"
        typePhone:
          type: string
          minLength: 1
          maxLength: 32
          example: "MOBILE"
        status:
          $ref: '#/components/schemas/Status'
        created_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          example: "2026-05-01T16:35:00Z"
        updated_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          example: "2026-05-01T16:40:00Z"
      required:
        - id
        - phone
        - typePhone
        - status

    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          minLength: 3
          maxLength: 3
          pattern: '^[0-9]{3}$'
          example: "404"
        message:
          type: string
          minLength: 1
          maxLength: 255
          example: Resource not found.
        detail:
          type: string
          maxLength: 500
          example: No information was found for the submitted request.
      required:
        - code
        - message